// src/app/water/components/models/water.report.model.ts
export interface WaterReport {
  calcium: number | null;
  magnesium: number | null;
  sodium: number | null;
  chloride: number | null;
  sulfate: number | null;
  alkalinity: number | null;
  mashVolume: number| null;
  spargeVolume: number| null;
  mashRoPercentage: number| null;
  spargeRoPercentage: number| null;
}